fix: Banner Concurrent UI Pattern 적용 (#45)
