+++ src/components/loading/Loading.tsx
@@ -1,0 +1,9 @@
+const Loading = () => {
+  return (
+    <figure className="mx-auto inline-flex items-center justify-center p-5">
+      <img alt="loading" src="/images/loading.gif" className="w-10" />
+    </figure>
+  );
+};
+
+export default Loading;
