+++ src/components/subject/RecentlyUsedSubjectListSkeleton.tsx
@@ -1,11 +1,7 @@
+import Loading from '../loading/Loading';
+
 const RecentlyUsedSubjectListSkeleton = () => {
-  return (
-    <div
-      className={`relative flex aspect-square h-[110px] w-full flex-shrink-0 select-none items-center justify-center rounded-[8px] bg-neutral-light-4 p-[8px] hover:cursor-pointer`}
-    >
-      로딩중
-    </div>
-  );
+  return <Loading />;
 };
 
 export default RecentlyUsedSubjectListSkeleton;
